feat: add command to add an asset bundle to a manifest (#8)
Takes a directory of asset bundles and automatically computes the name,
hash, dependencies, size, and child assets for the manifest entry. Then
adds them to the manifest. This is based on the internal bundle asset at
PathId 1.

The only unknown is the 'group' parameter - I'm not clear on how this is
calculated or what it actually means.